Update deployment schedule
Deploy at 1800 UTC so blog is updated around 12/1 Central Time. Also,
set timezone to UTC, since deployments are scheduled using UTC time.

The build on 2020-04-15 0000 UTC ran successfully, but the blog entry
for that date was skipped because jekyll claimed the date was in the
future. This suggests to me that the default GitHub Action VM is not
running in UTC timeszone for whatever reason.

See https://github.com/kadler/kadler.github.io/runs/587429991?check_suite_focus=true#step:6:11